clean: Refactor analog compiler code
