+++ src/oqd_core/compiler/analog/operator/__init__.py
@@ -12,10 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .compile import compile_analog_circuit
-from .assign import infer_analog_circuit_dim_cfg
-
-__all__ = [
-    "compile_analog_circuit",
-    "infer_analog_circuit_dim_cfg",
-]
