+++ src/oqd_core/compiler/analog/passes/assign.py
@@ -15,8 +15,8 @@
 ########################################################################################
 from oqd_core.interface.analog import Evolve
 from oqd_core.compiler.analog.error import AnalogCompilerError
-from oqd_core.compiler.analog.operator_dim import operator_dim
-from oqd_core.compiler.analog.cfg_walk import iter_stmt_blocks
+from oqd_core.compiler.analog.operator.dim import operator_dim
+from oqd_core.compiler.analog.cfg.walk import iter_stmt_blocks
 from oqd_core.analysis.utils.control_flow import ControlFlowGraph
