+++ src/oqd_core/compiler/analog/passes/compile.py
@@ -13,10 +13,10 @@
 # limitations under the License.
 
 from oqd_core.backend.metric import Expectation
-from oqd_core.compiler.analog.cfg_walk import canonicalize_math_cfg
-from oqd_core.compiler.analog.operator_env import canonicalize_operators_cfg
+from oqd_core.compiler.analog.cfg.walk import canonicalize_math_cfg
+from oqd_core.compiler.analog.cfg.operator_env import canonicalize_operators_cfg
 from oqd_core.compiler.analog.passes.assign import infer_analog_circuit_dim_cfg
-from oqd_core.compiler.analog.passes.canonicalize import analog_operator_canonicalization
+from oqd_core.compiler.analog.operator.canonicalize import canonicalize_operator_expr
 from oqd_core.compiler.analog.verify.passes import verify_analog_args_dim, verify_hamiltonian_target_dim, verify_register_access_dim
 from oqd_core.analysis.analog.analyze import AnalogAnalysisResult
 from oqd_core.interface.analog import AnalogCircuit
@@ -31,7 +31,7 @@ __all__ = [ "compile_analog_circuit" ]
 def canonicalize_args_metrics(args):
     for metric in args.metrics.values():
         if isinstance(metric, Expectation):
-            metric.operator = analog_operator_canonicalization(metric.operator)
+            metric.operator = canonicalize_operator_expr(metric.operator)
 
 
 def compile_analog_circuit(model: AnalogCircuit, analysis_result: AnalogAnalysisResult,  args=None):
