+++ src/oqd_core/compiler/analog/verify/passes.py
@@ -12,13 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from oqd_core.compiler.analog.operator_dim import operator_dim
+from oqd_core.compiler.analog.operator.dim import operator_dim
 from oqd_core.analysis.analog.symbol_table import AnalogSymbolTable, target_dim
 from oqd_core.analysis.utils.control_flow import ControlFlowGraph
 from oqd_core.interface.analog import Evolve, Initialize, Measure
 from oqd_core.backend.metric import Expectation
 from oqd_core.compiler.analog.error import AnalogCompilerError
-from oqd_core.compiler.analog.cfg_walk import iter_stmt_blocks
+from oqd_core.compiler.analog.cfg.walk import iter_stmt_blocks
 
 __all__ = [
     "verify_analog_args_dim",
